requirements.txt
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
google-api-core==1.25.1
google-auth==1.25.0
google-cloud-core==1.5.0
google-cloud-storage==1.35.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
idna==2.10
numpy==1.20.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pytz==2021.1
veryfi
requests==2.25.1
rsa==4.7
six==1.15.0
urllib3==1.26.3
veryfi==0.0.5
